Port to material design
progress left

+++ WorkoutManager.Contract/Models/ExerciseSet/IExerciseSet.cs
@@ -2,6 +2,6 @@
 {
     public interface IExerciseSet
     {
-        double Weight { get; set; }
+        int Weight { get; set; }
     }
 }
