+++ WorkoutManager.Contract/Models/ExerciseSet/IsometricExerciseSet.cs
@@ -7,7 +7,7 @@ namespace WorkoutManager.Contract.Models.ExerciseSet
     {
         public int Id { get; set; }
         
-        public double Weight { get; set; }
+        public int Weight { get; set; }
         
         public TimeSpan Duration { get; set; }
